Refactor useRunningViceJobs hook to utils with refresh
