+++ src/components/notifications/RunningViceTab.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 
 import { build } from "@cyverse-de/ui-lib";
 import { List, ListItem, Typography } from "@material-ui/core";
@@ -6,26 +6,12 @@ import { Skeleton } from "@material-ui/lab";
 
 import { openInteractiveUrl } from "components/analyses/utils";
 import DELink from "components/utils/DELink";
-import { useUserProfile } from "contexts/userProfile";
 import { useTranslation } from "i18n";
 import ids from "./ids";
-import { useRunningViceJobs } from "serviceFacades/analyses";
 
 function RunningViceTab(props) {
-    const { baseId, handleClose } = props;
+    const { baseId, handleClose, analyses, isFetching } = props;
     const { t } = useTranslation("common");
-    const [analyses, setAnalyses] = useState([]);
-    const [userProfile] = useUserProfile();
-
-    const { isFetching } = useRunningViceJobs({
-        enabled: userProfile?.id,
-        onSuccess: (resp) => {
-            const analyses = resp?.analyses?.filter(
-                (analysis) => analysis?.interactive_urls?.length > 0
-            );
-            setAnalyses(analyses);
-        },
-    });
 
     if (isFetching) {
         return (
