+++ src/serviceFacades/analyses.js
@@ -1,9 +1,6 @@
 import callApi from "../common/callApi";
 
 import AnalysisStatus from "components/models/analysisStatus";
-import { useQuery } from "react-query";
-import analysisStatus from "components/models/analysisStatus";
-import appType from "../components/models/AppType";
 
 const ANALYSES_LISTING_QUERY_KEY = "fetchAnalysesListingKey";
 const ANALYSIS_HISTORY_QUERY_KEY = "fetchAnalysisHistoryKey";
@@ -34,27 +31,6 @@ function getAnalyses(key, { rowsPerPage, orderBy, order, page, filter }) {
         endpoint: "/api/analyses",
         method: "GET",
         params,
-    });
-}
-
-const runningViceJobsFilter = [
-    { field: "status", value: analysisStatus.RUNNING },
-    { field: "type", value: appType.interactive.value },
-    { field: "ownership", value: "mine" },
-];
-
-function useRunningViceJobs({ enabled, onSuccess, onError }) {
-    return useQuery({
-        queryKey: [
-            ANALYSES_LISTING_QUERY_KEY,
-            { filter: runningViceJobsFilter },
-        ],
-        queryFn: getAnalyses,
-        config: {
-            enabled,
-            onSuccess,
-            onError,
-        },
     });
 }
 
@@ -118,6 +94,7 @@ function getAnalysisRelaunchInfo(key, { id }) {
         method: "GET",
     });
 }
+
 /**
  * Search Analyses
  * @param {string} key - react-query key
@@ -186,7 +163,6 @@ export {
     submitAnalysis,
     updateAnalysisComment,
     searchAnalysesInfinite,
-    useRunningViceJobs,
     ANALYSES_LISTING_QUERY_KEY,
     ANALYSIS_HISTORY_QUERY_KEY,
     ANALYSIS_PARAMS_QUERY_KEY,
